main_service/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
djangorestframework==3.15.1
gunicorn==21.2.0
idna==3.7
packaging==24.0
redis==5.0.4
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
